Revert PaymentMethodVerticalLayoutUITest refactor
